feat: Issue #59 Add support for advantage/disadvantage
If you shift-click a skill a dialog where you can choose advantage/disadvantage, difficulty and modifiers show up.
Note that this skill roll dialog uses the Cepheus Engine rules, i.e. it adds a modifier for difficulty rather than change the target number. A roll above 8 is considered a success. The Effect of the roll is the result - 8.
I'd like to make the result row show 'Effect: 0' (or whatever the result was), but haven't figured out how yet, so that will have to wait for another time.