feat: add expandable roll descriptions (#1185)
This update adds an expandable table to chat rolls that describes roll calculations. Click on the "Rolling:" text to show the table.  Also added is an icon for the skill rolled (if relevant).  The 'Rolling' text no longer shows the modifiers by default.  For example, now displays just 'STR' rather than 'STR(+2)'.  There is a new display option to show these values, if desired.