feat: Allow user defined armor handling (#1294)
This update allows the user to specify more complex formulas when applying damage to an actor.  Currently, the default is, (damage - effective armor) = damage applied.  You can now specify the formula to apply damage in the "Armor - damage reduction formula" field under the rules settings.  The formula's default is set as "@damage - @effectiveArmor", which is the way it has worked up to now and is the formula for almost all Cepheus / Traveller systems.

The armor-damage formula can be any valid roll formula and can use the TARGET's system data.  For example, you could use an alternate damage formula to also reduce the damage by the target's dexterity mod by using the formula: "@damage - @effectiveArmor - @characteristics.dexterity.mod".  However, one caution is not to use any random (dice) terms in the formula as the results will be unpredictable.

A big thanks to @Kampfzwerg42 who authored this release.